app/requirements.txt
requests
psycopg2-binary
polyline
shapely
flask
pandas
gevent
python-dotenv
geopy

